move all UI stuff to the sample project
