+++ libnetcipher/src/info/guardianproject/onionkit/ui/OrbotHelper.java
@@ -2,17 +2,13 @@
 package info.guardianproject.onionkit.ui;
 
 import android.app.Activity;
-import android.app.AlertDialog;
 import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.net.Uri;
 import android.text.TextUtils;
 import android.util.Log;
-
-import info.guardianproject.onionkit.R;
 
 import java.util.List;
 
@@ -65,60 +61,6 @@ public class OrbotHelper {
         } catch (PackageManager.NameNotFoundException e) {
             return false;
         }
-    }
-
-    /**
-     * Ask the user whether to install Orbot or not. Check if installing from
-     * F-Droid or Google Play, otherwise take the user to the Orbot download
-     * page on f-droid.org.
-     */
-    public static void promptToInstall(final Activity activity) {
-        String message = activity.getString(R.string.you_must_have_orbot) + "  ";
-        activity.getString(R.string.get_orbot_from_google_play);
-
-        final Intent intent = getOrbotInstallIntent(activity);
-        if (intent.getPackage() == null) {
-            message += activity.getString(R.string.download_orbot_from_fdroid);
-        } else {
-            message += activity.getString(R.string.get_orbot_from_fdroid);
-        }
-
-        AlertDialog.Builder builder = new AlertDialog.Builder(activity);
-        builder.setTitle(R.string.install_orbot_);
-        builder.setMessage(message);
-        builder.setPositiveButton(R.string.yes, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                activity.startActivity(intent);
-            }
-        });
-        builder.setNegativeButton(R.string.no, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                // nothing to do
-            }
-        });
-        builder.show();
-    }
-
-    public static void requestOrbotStart(final Activity activity) {
-        AlertDialog.Builder downloadDialog = new AlertDialog.Builder(activity);
-        downloadDialog.setTitle(R.string.start_orbot_);
-        downloadDialog
-                .setMessage(R.string.orbot_doesn_t_appear_to_be_running_would_you_like_to_start_it_up_and_connect_to_tor_);
-        downloadDialog.setPositiveButton(R.string.yes, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-                activity.startActivityForResult(getOrbotStartIntent(), 1);
-            }
-        });
-        downloadDialog.setNegativeButton(R.string.no, new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialogInterface, int i) {
-            }
-        });
-        downloadDialog.show();
-
     }
 
     public static void requestHiddenServiceOnPort(Activity activity, int port) {
